Fix integer types

+++ demos/device_defender_for_aws/metrics_collector/stub/metrics_collector.c
@@ -89,8 +89,8 @@ MetricsCollectorStatus_t GetNetworkStats( NetworkStats_t * pOutNetworkStats )
 /*-----------------------------------------------------------*/
 
 MetricsCollectorStatus_t GetOpenTcpPorts( uint16_t * pOutTcpPortsArray,
-                                          uint32_t tcpPortsArrayLength,
-                                          uint32_t * pOutNumTcpOpenPorts )
+                                          size_t tcpPortsArrayLength,
+                                          size_t * pOutNumTcpOpenPorts )
 {
     MetricsCollectorStatus_t status = MetricsCollectorSuccess;
 
@@ -126,8 +126,8 @@ MetricsCollectorStatus_t GetOpenTcpPorts( uint16_t * pOutTcpPortsArray,
 /*-----------------------------------------------------------*/
 
 MetricsCollectorStatus_t GetOpenUdpPorts( uint16_t * pOutUdpPortsArray,
-                                          uint32_t udpPortsArrayLength,
-                                          uint32_t * pOutNumUdpOpenPorts )
+                                          size_t udpPortsArrayLength,
+                                          size_t * pOutNumUdpOpenPorts )
 {
     MetricsCollectorStatus_t status = MetricsCollectorSuccess;
 
@@ -163,8 +163,8 @@ MetricsCollectorStatus_t GetOpenUdpPorts( uint16_t * pOutUdpPortsArray,
 /*-----------------------------------------------------------*/
 
 MetricsCollectorStatus_t GetEstablishedConnections( Connection_t * pOutConnectionsArray,
-                                                    uint32_t connectionsArrayLength,
-                                                    uint32_t * pOutNumEstablishedConnections )
+                                                    size_t connectionsArrayLength,
+                                                    size_t * pOutNumEstablishedConnections )
 {
     MetricsCollectorStatus_t status = MetricsCollectorSuccess;
